Revert "Added atomictorrent repo, and added multirepo support for pragmo:"
This reverts commit a4873c43c2d49ddfc02143a38257e774c5f578a2, but undoes
the revert for all the documented changes above.  The only change that
stays reverted is the (undocumented) change to the tinydns autogen script.

If you want the tinydns autogen script to output your own NS instead
of mine (the default), the correct thing to do is to change your own
configuration (for example, by running ./configure).  Starting an editing
war with everybody else who's too lazy to configure and decides to pull
the same stunt you did doesn't help anyone, since the whole lot of you
will be constantly stepping all over each other.  I believe my NS is
the most sane default, so I think it should stay.  If I'm wrong, feel
free to revert my revert of your commit, but if you're going to do a
commit that's likely to cause controversy, it'd be nice to explain your
reasons in the commit message, rather than just trying to slip it under
the radar by bundling it in with another commit and not even mentioning
it in the log.  (Remember: I can't read your mind, so I don't think
it's fair for you to expect me to know you're right, even if you are.
I'm just not smart enough to know why you're right.)